+++ .github/workflows/integration-test.yml
@@ -96,7 +96,7 @@ jobs:
         run: |
           echo "check if opensearch is ready"
           curl -XGET https://localhost:9200 -u 'admin:admin' -k
-          yarn test:jest_server --coverage
+          node ./test/run_jest_tests.js --config ./test/jest.config.server.js
         working-directory: ${{ steps.install-dashboards.outputs.plugin-directory }}
 
       #- name: Run integration tests on Windows
