+++ test/jest_integration/runIdpServer.js
@@ -13,20 +13,20 @@
  *   permissions and limitations under the License.
  */
 
-const { runServer } = require('saml-idp');
-
-const { generate } = require('selfsigned');
-
-const pems = generate(null, {
-  keySize: 2048,
-  clientCertificateCN: '/C=US/ST=California/L=San Francisco/O=JankyCo/CN=Test Identity Provider',
-  days: 7300,
-});
-
-// Create certificate pair on the fly and pass it to runServer
-runServer({
-  acsUrl: 'http://localhost:5601/_opendistro/_security/saml/acs',
-  audience: 'https://localhost:9200',
-  cert: pems.cert,
-  key: pems.private.toString().replace(/\r\n/, '\n'),
-});
+//const { runServer } = require('saml-idp');
+//
+//const { generate } = require('selfsigned');
+//
+//const pems = generate(null, {
+//  keySize: 2048,
+//  clientCertificateCN: '/C=US/ST=California/L=San Francisco/O=JankyCo/CN=Test Identity Provider',
+//  days: 7300,
+//});
+//
+//// Create certificate pair on the fly and pass it to runServer
+//runServer({
+//  acsUrl: 'http://localhost:5601/_opendistro/_security/saml/acs',
+//  audience: 'https://localhost:9200',
+//  cert: pems.cert,
+//  key: pems.private.toString().replace(/\r\n/, '\n'),
+//});
